+++ nncf/onnx/hardware/config.py
@@ -16,7 +16,7 @@ from typing import Type
 
 from nncf.common.graph import OperatorMetatype
 from nncf.common.hardware.config import HWConfig
-from nncf.experimental.onnx.graph.metatypes.onnx_metatypes import get_operator_metatypes
+from nncf.onnx.graph.metatypes.onnx_metatypes import get_operator_metatypes
 
 
 class ONNXHWConfig(HWConfig):
