+++ nncf/onnx/statistics/collectors.py
@@ -22,10 +22,10 @@ from nncf.common.tensor_statistics.collectors import MinMaxStatisticCollector
 from nncf.common.tensor_statistics.collectors import NNCFCollectorTensorProcessor
 from nncf.common.tensor_statistics.collectors import MeanMinMaxStatisticCollector
 from nncf.common.tensor_statistics.collectors import MeanStatisticCollector
-from nncf.experimental.onnx.tensor import ONNXNNCFTensor
-from nncf.experimental.onnx.statistics.statistics import ONNXMinMaxTensorStatistic
-from nncf.experimental.onnx.statistics.statistics import ONNXMeanTensorStatistic
-from nncf.experimental.onnx.statistics.statistics import ONNXBatchTensorStatistic
+from nncf.onnx.tensor import ONNXNNCFTensor
+from nncf.onnx.statistics.statistics import ONNXMinMaxTensorStatistic
+from nncf.onnx.statistics.statistics import ONNXMeanTensorStatistic
+from nncf.onnx.statistics.statistics import ONNXBatchTensorStatistic
 
 
 class ONNXNNCFCollectorTensorProcessor(NNCFCollectorTensorProcessor):
