+++ nncf/quantization/algorithms/post_training/algorithm.py
@@ -20,7 +20,6 @@ from nncf.parameters import TargetDevice
 from nncf.common.quantization.structs import QuantizationPreset
 from nncf.common.utils.backend import BackendType
 from nncf.common.utils.backend import get_backend
-from nncf.common.logging import nncf_logger
 from nncf.quantization.algorithms.algorithm import Algorithm
 from nncf.quantization.algorithms.algorithm import AlgorithmParameters
 from nncf.quantization.algorithms.definitions import RangeType
@@ -170,7 +169,7 @@ class PostTrainingQuantization(Algorithm):
         :return: backnd-specific StatisticsAggregator
         """
         if backend == BackendType.ONNX:
-            from nncf.experimental.onnx.statistics.aggregator import \
+            from nncf.onnx.statistics.aggregator import \
                 ONNXStatisticsAggregator
             return ONNXStatisticsAggregator(dataset)
         return None
@@ -183,10 +182,6 @@ class PostTrainingQuantization(Algorithm):
         modified_model = deepcopy(model)
         if statistic_points is None:
             backend = get_backend(modified_model)
-
-            # TODO (KodiaqQ): Remove after ONNX is removed from experimental
-            if backend == BackendType.ONNX:
-                nncf_logger.warning('You are using the experimental ONNX backend for post-training quantization.')
 
             statistics_aggregator = self._create_statistics_aggregator(dataset, backend)
             for algorithm in self.algorithms:
