+++ tests/onnx/benchmarking/README.md
@@ -135,7 +135,7 @@ Please refer to [Docker image build](#docker-image-build) section.
                 -v <OUTPUT_DIR>:/output     \
                 onnx_ptq_experimental:dev
 
-(container) $ nncf/examples/experimental/onnx/run_ptq_onnx_models.sh [classification|object_detection_segmentation] /onnx-models /output /omz_data $NUMBER_OF_SAMPLES
+(container) $ nncf/tests/onnx/benchmarking/run_ptq_onnx_models.sh [classification|object_detection_segmentation] /onnx-models /output /omz_data $NUMBER_OF_SAMPLES
 ```
 
 You have to choose the model type between `[classification|det_and_seg]`.
@@ -145,13 +145,13 @@ For examples, to run with `NUMBER_OF_SAMPLES=500`, you can command as follows.
 1. Classification models
 
 ```bash
-(container) $ nncf/examples/experimental/onnx/run_ptq_onnx_models.sh classification /onnx-models /output /omz_data 500
+(container) $ nncf/tests/onnx/benchmarking/run_ptq_onnx_models.sh classification /onnx-models /output /omz_data 500
 ```
 
 2. Object detection and segmentation models
 
 ```bash
-(container) $ nncf/examples/run_ptq_onnx_models.sh object_detection_segmentation /onnx-models /output /omz_data 500
+(container) $ nncf/tests/onnx/benchmarking/run_ptq_onnx_models.sh object_detection_segmentation /onnx-models /output /omz_data 500
 ```
 
 After benchmark is done, outputs are located in `/output` which is a mounted directory from the host path `<OUTPUT_DIR>`.
